+++ lib/parse/stack.rb
@@ -11,13 +11,17 @@ require_relative 'webhooks'
 module Parse
   class Error < StandardError; end;
   module Stack
+
+  end
+
+  class Hyperdrive
     # Applies a remote JSON hash containing the ENV keys and values from a remote
     # URL. Values from the JSON hash are only applied to the current ENV hash ONLY if
     # it does not already have a value. Therefore local ENV values will take precedence
     # over remote ones. By default, it uses the url in environment value in 'CONFIG_URL' or 'HYPERDRIVE_URL'.
     # @param url [String] the remote url that responds with the JSON body.
     # @return [Boolean] true if the JSON hash was found and applied successfully.
-    def self.hyperdrive_config!(url = nil)
+    def self.config!(url = nil)
       url ||= ENV["HYPERDRIVE_URL"] || ENV['CONFIG_URL']
       if url.present?
         begin